Address review: extract shared helpers, add set feature gate
- Extract anna_binary, server_path, config_file, start_servers,
  stop_servers, ServerGuard into tests/common/mod.rs
- Both simple.rs and system.rs now use shared module
- Add #[cfg(feature = "set")] gate around put_set/get_set tests
  so cargo test --no-default-features compiles
